Parameter editor implementation

+++ crhmcode/src/modules/ClassCRHMCanopyClearingGap.cpp
@@ -517,7 +517,7 @@ void ClassCRHMCanopyClearingGap::run(void){
           double IceBulbT = hru_t[hh] - (Vi* Hs/1e6/ci);
           double Six_Hour_Divisor = Global::Freq/4.0; // used to unload over 6 hours
 
-          const float U = -1 * log(0.678) / (24 * 7 * Global::Freq / 24); // weekly dimensionless unloading coefficient -> to CRHM time interval
+          const double U = -1 * log(0.678) / (24 * 7 * Global::Freq / 24); // weekly dimensionless unloading coefficient -> to CRHM time interval
                                                        // 21Mar2022 correction: invert the term 24/Global::Freq, use unloading rate coefficient U = -log(c)/t for snow unloading determined by inverse function of c = e^(-Ut) = 0.678 based on Eq. 14 in Hedstrom and Pomeroy (1998)
   // determine whether canopy snow is unloaded:
